refactor: move cn to /app/lib
