+++ src/app/components/ui/label.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/lib/cn"
 
 function Label({
   className,
