+++ src/app/components/ui/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/lib/cn"
 
 function Textarea({ className, ...props }: React.ComponentProps<"textarea">) {
   return (
